Fix "squishy" animation and close-menu threshold
Having an animation block on the "changed" state of the gesture recognizer makes the panel drag feel somewhat squishy. Since the "end" state calls an animated transition, it's just fine to perform the changes without animation. Unfortunately things are still not natural because of the time-warp effect due to the content container scaling.